change fputb and fgetb to fputch and fgetch
The chars are returned an applied directly, with two extra functions,
inttochar and chartoint, exposed to allow working with chars.

This is taking the philosophy of not hiding the implementation details
of the target language interfaces, and relying a layer of Smooth code
to abstract these away.

At current, this type of decision will slow down the output code, but
in time, optimisations in Smooth will minimise or in some cases
reverse this slow-down.